requirement.txt
phonenumber
opencage
folium
geocoder 
ip2geotools